safari-600.7-branch/LayoutTests/inspector-obsolete/console/console-exception-stack-traces-expected.txt

CONSOLE MESSAGE: line 16: URIError: URI error
CONSOLE MESSAGE: line 21: TypeError: 1 is not a function (evaluating '1()')
CONSOLE MESSAGE: line 26: RangeError: Array size is not a small enough positive integer.
CONSOLE MESSAGE: line 31: SyntaxError: Expected token '}'
CONSOLE MESSAGE: line 36: ReferenceError: Can't find variable: doesNotExist
CONSOLE MESSAGE: line 46: Error: Custom Error
CONSOLE MESSAGE: line 51: [object Object]
CONSOLE MESSAGE: line 56: error string
CONSOLE MESSAGE: line 1: DONE
Tests that JavaScript exception's contain stack traces and that the functions match console.trace() stack traces.

SUCCESS
